Adddress code review comments
