+++ common/include/common_host_helpers.hpp
@@ -5,6 +5,7 @@
 #pragma once
 
 #include <fstream>
+#include <iostream>
 #include <string>
 
 #include <fmt/core.h>
